+++ lib/puppetserver/ca/action/revoke.rb
