+++ lib/puppetserver/ca/cli.rb
@@ -5,7 +5,7 @@ require 'puppetserver/ca/action/clean'
 require 'puppetserver/ca/action/create'
 require 'puppetserver/ca/action/import'
 require 'puppetserver/ca/action/generate'
-require 'puppetserver/ca/revoke_action'
+require 'puppetserver/ca/action/revoke'
 require 'puppetserver/ca/action/list'
 require 'puppetserver/ca/sign_action'
 require 'puppetserver/ca/utils'
